+++ demos/manual/src/main/java/is/codion/framework/demos/manual/common/demo/ApplicationPanel.java
@@ -325,7 +325,7 @@ public final class ApplicationPanel extends JPanel {
 										.focusable(false)
 										.border(createTitledBorder("Message"))
 										.enabled(inputEnabledState)
-										.linkedValue(model.message())
+										.link(model.message())
 										.build(component -> add(component, BorderLayout.SOUTH)))
 						.add(lookAndFeelComboBox(true))
 						.build(southPanel -> add(southPanel, BorderLayout.SOUTH));