+++ framework/domain/src/test/java/is/codion/framework/domain/entity/DefaultEntityDefinitionTest.java
@@ -185,8 +185,8 @@ public class DefaultEntityDefinitionTest {
 								entityType.integerColumn("id").define().primaryKey(),
 								name.define().column(),
 								info.define().column(),
-								derived.define().derived(linkedValues ->
-												linkedValues.get(name).toString() + linkedValues.get(info), name, info)));
+								derived.define().derived(sourceValues ->
+												sourceValues.get(name).toString() + sourceValues.get(info), name, info)));
 			}
 		}
 		Domain domain = new TestDomain();
@@ -304,7 +304,7 @@ public class DefaultEntityDefinitionTest {
 								pk.define().primaryKey(),
 								column1.define().column(),
 								column2.define().column(),
-								der.define().derived(linkedValues -> null, column1, column2)));
+								der.define().derived(sourceValues -> null, column1, column2)));
 			}
 		}
 		Domain domain = new TestDomain();