+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/label/DefaultLabelBuilder.java
@@ -52,7 +52,7 @@ final class DefaultLabelBuilder<T> extends AbstractComponentBuilder<T, JLabel, L
 
 	DefaultLabelBuilder(ValueObserver<T> linkedValue) {
 		this.text = null;
-		linkedValue(requireNonNull(linkedValue));
+		link(requireNonNull(linkedValue));
 	}
 
 	@Override